change Shim layer's name to spark311cdh from spark311cloud
