--- shims/spark311cloud/pom.xml
@@ -1,101 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Copyright (c) 2020-2021, NVIDIA CORPORATION.
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.nvidia</groupId>
-        <artifactId>rapids-4-spark-shims_2.12</artifactId>
-        <version>0.6.0-SNAPSHOT</version>
-	<relativePath>../pom.xml</relativePath>
-    </parent>
-    <artifactId>rapids-4-spark-shims-spark311cloud_2.12</artifactId>
-    <name>RAPIDS Accelerator for Apache Spark SQL Plugin Spark 3.1.1 Cloud Shim</name>
-    <description>The RAPIDS SQL plugin for Apache Spark 3.1.1 Shim</description>
-    <version>0.6.0-SNAPSHOT</version>
-
-    <!-- Set 'spark.version' for the shims layer -->
-    <!-- Create a separate file 'SPARK_VER.properties' in the jar to save cudf & spark version info -->
-    <build>
-        <plugins>
-            <plugin>
-                <artifactId>maven-antrun-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>dependency</id>
-                        <phase>generate-resources</phase>
-                        <configuration>
-                            <target>
-                                <mkdir dir="${project.build.directory}/extra-resources"/>
-                                <exec executable="bash" failonerror="true" output="${project.build.directory}/extra-resources/spark-${spark311cloud.version}-info.properties">
-                                    <arg value="${user.dir}/build/dependency-info.sh"/>
-                                    <arg value="${cudf.version}"/>
-                                    <arg value="${cuda.version}"/>
-                                    <arg value="${spark311cloud.version}"/>
-                                </exec>
-                            </target>
-                        </configuration>
-                        <goals>
-                            <goal>run</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.scalastyle</groupId>
-                <artifactId>scalastyle-maven-plugin</artifactId>
-            </plugin>
-        </plugins>
-
-        <resources>
-          <resource>
-            <!-- Include the properties file to provide the build information. -->
-            <directory>${project.build.directory}/extra-resources</directory>
-          </resource>
-          <resource>
-            <directory>src/main/resources</directory>
-          </resource>
-        </resources>
-    </build>
-
-    <dependencies>
-	    <dependency>
-            <groupId>com.nvidia</groupId>
-            <artifactId>rapids-4-spark-shims-spark311_${scala.binary.version}</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.spark</groupId>
-            <artifactId>spark-sql_${scala.binary.version}</artifactId>
-            <version>${spark311cloud.version}</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.apache.curator</groupId>
-                    <artifactId>curator-recipes</artifactId>
-                </exclusion>
-            </exclusions>
-	        <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.curator</groupId>
-            <artifactId>curator-recipes</artifactId>
-            <version>4.3.0.7.2.7.0-184</version>
-        </dependency>
-    </dependencies>
-</project>
