--- shims/spark311cloud/src/main/resources/META-INF/services/com.nvidia.spark.rapids.SparkShimServiceProvider
@@ -1,1 +1,0 @@
-com.nvidia.spark.rapids.shims.spark311cloud.SparkShimServiceProvider
